Cart implementation

+++ db/src/main/java/ru/alekseiadamov/db/dto/ProductDTO.java
@@ -10,6 +10,7 @@ import ru.alekseiadamov.db.entity.Category;
 import javax.validation.constraints.DecimalMin;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -24,7 +25,7 @@ public class ProductDTO {
 
     @NotNull
     @DecimalMin(value = "0.01")
-    private Double price;
+    private BigDecimal price;
 
     @NotNull
     private Category category;
@@ -36,7 +37,7 @@ public class ProductDTO {
 
     private MultipartFile[] newPictures;
 
-    public ProductDTO(Long id, String name, Double price, Category category, Brand brand) {
+    public ProductDTO(Long id, String name, BigDecimal price, Category category, Brand brand) {
         this.id = id;
         this.name = name;
         this.price = price;
